Fix crash / boot loop issue on ESP32-C3 (and probably many similar devices)
Symptoms:
- The device resets very early during boot, often right after the log line:
  - `Setting up Ferraris Meter...`
- Serial log shows a Guru Meditation crash:
  - `Guru Meditation Error: Core 0 panic'ed (Load access fault)`

Cause:
In `FerrarisMeter`, the member pointer `m_debounce_threshold_number` (type `number::Number*`) was not initialized in the constructor.

Fix:
The fix is to explicitly initialize `m_debounce_threshold_number` to `nullptr` in the constructor initializer list.

A second issue was also fixed: Logging used the wrong format specifier for a `uint64_t` rotation counter (`%u` instead of `%llu`). On ESP32 targets, `printf`-style format mismatches are undefined behavior and can corrupt the stack.

+++ components/ferraris/ferraris_meter.cpp
@@ -449,7 +449,7 @@ namespace esphome::ferraris
     void FerrarisMeter::set_rotation_counter(uint64_t value)
     {
         m_rotation_counter = value;
-        ESP_LOGD(TAG, "Set rotation counter:  %llu rotations", (unsigned long long) m_rotation_counter);
+        ESP_LOGD(TAG, "Set rotation counter:  %llu rotations", m_rotation_counter);
 
         update_energy_counter();
     }
